Move handler to composable
to keep the logic related to groups
loading in one place
